Updating validation signature
